backend/requirements.txt
fastapi
uvicorn
pydantic
kafka-python
elasticsearch
psycopg2-binary
aiokafka
pytest